put DifferentialUseAnalysis inside a namespace
